Refactor src folder structure (separate components and pages) and add react-dom-router package
