+++ apps/frontend/src/components/Card.jsx
