+++ apps/frontend/src/components/Navbar.jsx
